--- src/main/java/kr/hhplus/be/server/order/adapter/gateway/OrderRepositoryCustom.java
@@ -1,4 +1,0 @@
-package kr.hhplus.be.server.order.adapter.gateway;
-
-public interface OrderRepositoryCustom {
-}
