--- src/main/java/kr/hhplus/be/server/order/adapter/gateway/OrderRepositoryCustomImpl.java
@@ -1,13 +1,0 @@
-package kr.hhplus.be.server.order.adapter.gateway;
-
-import jakarta.persistence.EntityManager;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class OrderRepositoryCustomImpl implements OrderRepositoryCustom {
-
-    private final EntityManager em;
-
-}
