+++ src/main/java/kr/hhplus/be/server/order/application/usecase/command/PlaceOrderCommand.java
@@ -1,0 +1,12 @@
+package kr.hhplus.be.server.order.application.usecase.command;
+
+import org.springframework.lang.Nullable;
+
+import java.util.List;
+
+public record PlaceOrderCommand(
+        Long userId,
+        @Nullable Long userCouponId,
+        List<ProductItemCommand> items
+) {
+}
