+++ src/main/java/kr/hhplus/be/server/order/application/usecase/command/ProductItemCommand.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.order.application.usecase.command;
+
+public record ProductItemCommand(
+        Long productId,
+        int quantity
+) {
+}
