+++ src/main/java/kr/hhplus/be/server/order/infra/gateway/OrderItemJpaRepository.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.order.infra.gateway;
+
+import kr.hhplus.be.server.order.infra.gateway.entity.OrderItemEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderItemJpaRepository extends JpaRepository<OrderItemEntity, Long> {
+}
