+++ src/main/java/kr/hhplus/be/server/order/infra/gateway/OrderJpaRepository.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.order.infra.gateway;
+
+import kr.hhplus.be.server.order.infra.gateway.entity.OrderEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderJpaRepository extends JpaRepository<OrderEntity, Long>, OrderRepositoryCustom {
+}
