+++ src/main/java/kr/hhplus/be/server/order/infra/gateway/OrderRepositoryCustom.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.order.infra.gateway;
+
+public interface OrderRepositoryCustom {
+}
