+++ src/main/java/kr/hhplus/be/server/order/infra/gateway/OrderRepositoryCustomImpl.java
@@ -1,0 +1,13 @@
+package kr.hhplus.be.server.order.infra.gateway;
+
+import jakarta.persistence.EntityManager;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class OrderRepositoryCustomImpl implements OrderRepositoryCustom {
+
+    private final EntityManager em;
+
+}
