+++ src/main/java/kr/hhplus/be/server/order/infra/gateway/entity/OrderEntity.java
@@ -1,0 +1,72 @@
+package kr.hhplus.be.server.order.infra.gateway.entity;
+
+import jakarta.persistence.*;
+import kr.hhplus.be.server.order.domain.Order;
+import kr.hhplus.be.server.order.domain.enums.OrderStatus;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@Setter
+@Table
+public class OrderEntity {
+
+    @Id @GeneratedValue
+    @Column(name = "order_id")
+    private Long id;
+    private Long userId;
+
+    @Column(nullable = true)
+    private Long userCouponId;
+    private OrderStatus orderStatus;
+    private LocalDateTime createdAt;
+    private LocalDateTime updatedAt;
+
+    protected OrderEntity() {}
+
+    public OrderEntity(Long userId, Long userCouponId, OrderStatus orderStatus, LocalDateTime createdAt, LocalDateTime updatedAt) {
+        this.userId = userId;
+        this.userCouponId = userCouponId;
+        this.orderStatus = orderStatus;
+        this.createdAt = createdAt;
+        this.updatedAt = updatedAt;
+    }
+
+    // update를 위한 생성자
+    public OrderEntity(long id, long userId, long userCouponId, OrderStatus status, LocalDateTime createdAt, LocalDateTime updatedAt) {
+        this.id = id;
+        this.userId = userId;
+        this.userCouponId = userCouponId;
+        this.orderStatus = status;
+        this.createdAt = createdAt;
+        this.updatedAt = updatedAt;
+    }
+
+    public static OrderEntity fromDomain(Order order) {
+        return new OrderEntity(
+                order.getUserId(),
+                order.getUserCouponId(),
+                order.getStatus(),
+                order.getCreatedAt(),
+                order.getUpdatedAt()
+        );
+    }
+
+    public Order toDomain() {
+        return Order.create(userId, userCouponId, orderStatus, createdAt, updatedAt).assignId(id);
+    }
+
+    public static OrderEntity fromExistingDomain(Order order) {
+        return new OrderEntity(
+                order.getId(),
+                order.getUserId(),
+                order.getUserCouponId(),
+                order.getStatus(),
+                order.getCreatedAt(),
+                order.getUpdatedAt()
+        );
+    }
+}
