--- src/main/java/kr/hhplus/be/server/order/usecase/in/PlaceOrderCommand.java
@@ -1,12 +1,0 @@
-package kr.hhplus.be.server.order.usecase.in;
-
-import org.springframework.lang.Nullable;
-
-import java.util.List;
-
-public record PlaceOrderCommand(
-        Long userId,
-        @Nullable Long userCouponId,
-        List<ProductItemCommand> items
-) {
-}
