--- src/main/java/kr/hhplus/be/server/order/usecase/in/PlaceOrderInput.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.order.usecase.in;
-
-import kr.hhplus.be.server.order.usecase.out.PlaceOrderOutput;
-
-public interface PlaceOrderInput {
-    void orderItemCommand(PlaceOrderCommand placeOrderCommand, PlaceOrderOutput placeOrderOutput);
-}
