--- src/main/java/kr/hhplus/be/server/order/usecase/in/ProductItemCommand.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.order.usecase.in;
-
-public record ProductItemCommand(
-        Long productId,
-        int quantity
-) {
-}
